third_party/blink/web_tests/virtual/pna-iframes-enabled/external/wpt/fetch/private-network-access/iframe.tentative.https.window_include=from-private-expected.txt

This is a testharness.js-based test.
PASS private to local: failed preflight.
PASS private to local: missing CORS headers.
PASS private to local: missing PNA header.
FAIL private to local: success. assert_equals: expected "loaded" but got "timeout"
PASS private to private: no preflight required.
PASS private to public: no preflight required.
Harness: the test ran to completion.
